+++ src/main/java/nc/model/ModelCuboid.java
@@ -8,7 +8,7 @@ public class ModelCuboid extends ModelRenderer {
 	public ModelCuboid(ModelBase model, int textureWidth, int textureHeight, int textureX, int textureY, int width, int height, int depth, float posX, float posY, float posZ, double rotationX, double rotationY, double rotationZ, boolean mirror) {
 		super(model, textureX, textureY);
 		addBox(0F, 0F, 0F, depth, height, width);
-		setRotationPoint(posX, -posY - (float)height, -posZ);
+		setRotationPoint(posX, -posY - height, -posZ);
 		setTextureSize(textureWidth, textureHeight);
 		this.mirror = mirror;
 		rotateAngleY = (float)(2D*Math.PI*(0.25D - rotationY));
