+++ src/main/java/nc/util/EnergyHelper.java
@@ -6,7 +6,7 @@ import nc.config.NCConfig;
 public class EnergyHelper {
 	
 	public static int getEUTier(double powerRF) {
-		double euPerTick = powerRF / (double) NCConfig.rf_per_eu;
+		double euPerTick = powerRF / NCConfig.rf_per_eu;
 		if (euPerTick <= 32D) return 1;
 		int maxTier = ModCheck.gregtechLoaded() ? 10 : 4;
 		return (int) Math.min(Math.ceil(((Math.log(euPerTick)/Math.log(2D)) - 3D) / 2D), maxTier);
